Rct135

+++ src/components/Pages/AdminEditPage/EditHome/EditUnidades/DeleteItemModal/DeleteItemModal.jsx
@@ -1,18 +1,19 @@
-import './DeleteItemModal.css'
 import React, {useState} from 'react'
-import {deleteCarItem, updateUnidadesData} from '../../../../../../services/ApiClient'
+import {Editor} from '@tinymce/tinymce-react'
+
+import {deleteUnitItem, updateUnidadesData} from '../../../../../../services/ApiClient'
+import InputWithLabel from '../../../../../Form/InputWithLabel/InputWithLabel'
+import {useFormState} from '../../../../../../hooks/useFormState'
 import InputFile from '../../../../../Form/InputFile/InputFile'
 import {app} from '../../../../../../services/firebase'
-import {useFormState} from '../../../../../../hooks/useFormState'
-import {Editor} from '@tinymce/tinymce-react'
-import InputWithLabel from '../../../../../Form/InputWithLabel/InputWithLabel'
 import Loader from '../../../../../Loader/Loader'
+import './DeleteItemModal.css'
 
 
 function DeleteItemModal({deleteItem, element, hideModal}) {
 
-    const [imageSuccess, setImageSuccess] = useState('')
     const [isDisabled, setIsDisabled] = useState(false)
+    const [message, setMessage] = useState('')
 
     const {state, onChange} = useFormState(
         {
@@ -25,7 +26,7 @@ function DeleteItemModal({deleteItem, element, hideModal}) {
             error: {
                 url: true,
                 desc: true,
-                logo: false,
+                logo: true,
             },
             touch: {},
         },
@@ -36,30 +37,35 @@ function DeleteItemModal({deleteItem, element, hideModal}) {
         }
     )
 
-    const {data} = state
+    const {data, error} = state
     const [registerError, setRegisterError] = useState(null)
 
     const [disabled, setDisabled] = useState(true)
 
     const deleteCarrouselItem = async (id) => {
-        const updateData = await deleteCarItem(id)
+        const updateData = await deleteUnitItem(id)
         deleteItem(updateData)
     }
 
     const editCarrouselItem = async (id) => {
 
-        try {
-            await updateUnidadesData(data, id)
-                .then(updateData => {
-                    deleteItem(updateData)
-                })
-        } catch (err) {
-            setRegisterError(err.response?.data?.message)
+        if (error.desc === false || error.logo === false) {
+            try {
+                await updateUnidadesData(data, id)
+                    .then(updateData => {
+                        deleteItem(updateData)
+                    })
+            } catch (err) {
+                setRegisterError(err.response?.data?.message)
+            }
+        } else {
+            setMessage('Por favor edite el logo o la descripción')
         }
     }
 
     const handleBannerDescription = (e) => {
         data.desc = e.target.getContent()
+        error.desc = false
     }
 
     const onFileSelected = async (e) => {
@@ -83,8 +89,8 @@ function DeleteItemModal({deleteItem, element, hideModal}) {
         // Get file url
         const fileUrl = await filePath.getDownloadURL()
         data.logo = fileUrl
-        setImageSuccess("Imagen subida correctamente")
         setIsDisabled(false)
+        error.logo = false
     }
 
     return (
@@ -95,14 +101,13 @@ function DeleteItemModal({deleteItem, element, hideModal}) {
                     <div className="row justify-content-center">
                         <div className="col-11 col-sm-5 EditElementsModal__container">
                             <span className="EditElementsModal__close" onClick={hideModal}></span>
-
                             <div className="col-sm-12">
                                 <p className="EditElementsModal__ask">Editar elemento {element.url}</p>
                                 <div className="card">
                                     <div className="card-body EditElementsModal__body">
                                         <div className="row align-items-center">
                                             <div className="col-sm-12">
-                                                <p className="EditElementsModal__text"><strong>Editar url</strong></p>
+                                                <p className="EditElementsModal__text"><strong>Url</strong></p>
                                             </div>
                                             <div className="col-sm-12">
                                                 <InputWithLabel
@@ -112,13 +117,14 @@ function DeleteItemModal({deleteItem, element, hideModal}) {
                                                     type="text"
                                                     cssStyle="form-control"
                                                     placeholder={element?.url}
+                                                    disabled
                                                 />
                                             </div>
                                             <div className="col-12">
                                                 <p className="EditElementsModal__text"><strong>Editar logo</strong></p>
                                             </div>
-                                            <div className="col-12 EditElementsModal__img">
-                                                <img src={element.logo} onerror="this.src = 'https://firebasestorage.googleapis.com/v0/b/grupo-leti-fd84e.appspot.com/o/images%2Fno-image.png?alt=media&token=73bf7cd8-629d-4deb-b281-9e629fbfb752';" alt={element.logo} />
+                                            <div className="col-12 EditElementsModal__img m-0">
+                                                <img src={element.logo} onError="this.src = 'https://firebasestorage.googleapis.com/v0/b/grupo-leti-fd84e.appspot.com/o/images%2Fno-image.png?alt=media&token=73bf7cd8-629d-4deb-b281-9e629fbfb752';" alt={element.logo} />
                                                 <InputFile
                                                     value={element?.logo}
                                                     onChange={onFileSelected}
@@ -127,6 +133,8 @@ function DeleteItemModal({deleteItem, element, hideModal}) {
                                                     type="file"
                                                     placeholder={element?.logo}
                                                 />
+                                            </div>
+                                            <div className="col-12">
                                             </div>
                                             <div className="col-12">
                                                 <p className="EditElementsModal__text"><strong>Editar descripción</strong></p>
@@ -149,10 +157,17 @@ function DeleteItemModal({deleteItem, element, hideModal}) {
                                                 />
                                             </div>
                                             <div className="col-12 col-sm-6">
-                                                <div onClick={() => editCarrouselItem(element._id)} className="leti-btn">Editar elemento</div>
+                                                <div onClick={() => deleteCarrouselItem(element._id)} className="leti-btn delete mt-0">Eliminar elemento</div>
                                             </div>
                                             <div className="col-12 col-sm-6">
-                                                <div onClick={() => deleteCarrouselItem(element._id)} className="leti-btn delete">Eliminar elemento</div></div>
+                                                <div onClick={() => editCarrouselItem(element._id)} className="leti-btn mt-0">Guardar cambios</div>
+                                            </div>
+                                            <div className="col-12 d-flex justify-content-end">
+                                                {message && <span className="AdminEdit__message">{message}</span>}
+                                            </div>
+                                            <div className="col-12">
+                                                {registerError && <div className="alert alert-danger">{registerError}</div>}
+                                            </div>
                                         </div>
                                     </div>
                                 </div>
